requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
cinemagoer==2023.5.1
greenlet==3.1.1
idna==3.10
IMDbPY==2022.7.9
lxml==5.3.1
requests==2.32.3
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
